+++ cucumber-spring/src/test/java/io/cucumber/spring/cucumbercontextconfigannotation/AbstractWithComponentAnnotation.java
@@ -1,0 +1,9 @@
+package io.cucumber.spring.cucumbercontextconfigannotation;
+
+import io.cucumber.spring.CucumberContextConfiguration;
+import org.springframework.test.context.ContextConfiguration;
+
+@CucumberContextConfiguration
+@ContextConfiguration("classpath:cucumber.xml")
+public abstract class AbstractWithComponentAnnotation {
+}